2D-3D Correspondence Pipeline (#14)
* modified dependencies

* included headers for finding correspondences

* superglue python files

* superglue weights

* calc_tools cpp files

* preprocess and camera cpp files

* main cpp files for finding correspondences and rosbag preprocessing

* added ransac capabilities (#13)

* thirdparty to deprepos changes

* moved superglue models to dep_repos

* integrated pipeline sequence in Calibration launch file

* launch file modifies

* change in launch file again

* removed velodyne driver

* aesthetic changes

* more aesthetic changes

* updated readme

* added gpu passthrough to devcontianer

* updated yaml and rviz configs

* updated readme and media

* updated readme

* moved superglue frm root to home in dockerfile

* made rviz default to true

* mdoifed superglue path

* updated rviz setup

* added dummy map link

* updated rviz setup

---------

+++ lidar_camera_calibration/src/calibration.cpp
@@ -73,14 +73,14 @@ private:
 
         cameraMatrix = cv::Mat::zeros(3, 3, CV_64F);
         
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             for (int j = 0; j < 3; ++j) {
                 file >> cameraMatrix.at<double>(i, j);
             }
         }
 
         Eigen::Matrix4d camera_matrix_eigen = Eigen::Matrix4d::Zero();
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             for (int j = 0; j < 3; ++j) {
                 camera_matrix_eigen(i, j) = cameraMatrix.at<double>(i, j);
             }
@@ -92,7 +92,7 @@ private:
     static std::string MatrixToString(const Eigen::Matrix4d& mat) {
         std::stringstream ss;
         ss << std::fixed << std::setprecision(4);
-        for (int i = 0; i < 4; ++i) {
+        for (int i = 0; i < (int)4; ++i) {
             for (int j = 0; j < 4; ++j) {
                 ss << std::setw(10) << mat(i,j) << " ";
             }
@@ -205,7 +205,7 @@ private:
         }
 
         file << "LiDAR Sensor to Camera Sensor Transformation Matrix:\n";
-        for (int i = 0; i < 4; ++i) {
+        for (int i = 0; i < (int)4; ++i) {
             for (int j = 0; j < 4; ++j) {
                 file << T_lidar_sensor_to_camera_sensor(i, j) << " ";
             }
@@ -213,7 +213,7 @@ private:
         }
 
         file << "\nCamera Matrix (K):\n";
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             for (int j = 0; j < 3; ++j) {
                 file << cameraMatrix.at<double>(i, j) << " ";
             }
@@ -295,7 +295,7 @@ private:
 
         points2D.clear();
         points3D.clear();
-        for (int i = 0; i < inliers.rows; ++i) {
+        for (int i = 0; i < (int)inliers.rows; ++i) {
             int idx = inliers.at<int>(i);
             points2D.push_back(temp_points2D[idx]);
             points3D.push_back(temp_points3D[idx]);
@@ -398,14 +398,14 @@ private:
 
         cv::Mat R_cv = cv::Mat(3, 3, CV_64F);
 
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             for (int j = 0; j < 3; ++j) {
                 R_cv.at<double>(i, j) = R(i, j);
             }
         }
 
         cv::Rodrigues(R_cv, rvec);
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             tvec.at<double>(i, 0) = t(i);
         }
 
@@ -433,7 +433,7 @@ private:
             cameraMatrix.at<double>(0, 2),  // cx
             cameraMatrix.at<double>(1, 2)   // cy
         };
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             rvex[i] = rvec.at<double>(i, 0);
             tvec_data[i] = tvec.at<double>(i, 0);
         }
@@ -469,7 +469,7 @@ private:
                     rvex[0], rvex[1], rvex[2]);
         RCLCPP_INFO(get_logger(), "Translation: [%f, %f, %f]", 
                     tvec_data[0], tvec_data[1], tvec_data[2]);
-        for (int i = 0; i < 3; ++i) {
+        for (int i = 0; i < (int)3; ++i) {
             rvec.at<double>(i, 0) = rvex[i];
             tvec.at<double>(i, 0) = tvec_data[i];
             axis_angle[i] = rvex[i];
